refactor(waveform-seek): H3.3 — split renderers into static + animated
